Range fixing 2

+++ Horner's Method/ValueCalculation.h
@@ -12,9 +12,9 @@ public:
 	static double HornerMethod(std::deque<double> &polynominal, int x);
 	static float HornerPrecision(std::deque<double> polynominal, int x_in, float desired_precision);
 	static float HornerPrecision(int degree, int x_in, float desired_precision);
-	static std::pair<int, int> HornerRange(std::deque<double>& polynominal, int x, int precision);
-	static double UpperFinder(std::deque<double>& polynominal, int i);
-	static double LowerFinder(std::deque<double>& polynominal, int i);
+	static std::pair<double, double> HornerRange(std::deque<double>& polynominal, int x, float precision);
+	static double UpperFinder(std::deque<double>& polynominal, double i);
+	static double LowerFinder(std::deque<double>& polynominal, double i);
 
 	static double TylorMethod(std::deque<double> &polynominal, int x, int n);
